// packages/core/src/db/db.d.ts
export interface GyomuAppsInfoCdtbl {
  description: string;
  id: string;
  mail_from_address: string | null;
  mail_from_name: string | null;
}

export interface GyomuMarketHoliday {
  holiday: string;
  id: string;
  market: string;
  year: number;
}

export interface GyomuMilestoneCdtbl {
  description: string;
  id: string;
  milestone_id: string;
}

export interface GyomuMilestoneDaily {
  id: string;
  milestone_id: string;
  modified_at: Date;
  modified_by: string;
  target_date: string;
  target_type: string;
  target_ym: string;
}

export interface GyomuParamMaster {
  id: string;
  item_fromdate: string | null;
  item_key: string;
  item_value: string;
}

export interface GyomuServiceCdtbl {
  description: string;
  id: string;
  parameter: string | null;
  service_type_id: string;
}

export interface GyomuServiceTypeCdtbl {
  assembly_name: string | null;
  class_name: string | null;
  description: string;
  id: string;
}

export interface GyomuStatusHandler {
  application_id: string;
  id: string;
  recipient_address: string | null;
  recipient_type: string | null;
  region: string | null;
  status_type_id: string | null;
}

export interface GyomuStatusInfo {
  application_id: string;
  description: string | null;
  developer_info: string | null;
  error_id: number;
  host_name: string | null;
  id: string;
  instance_id: number;
  modified_at: Date;
  modified_by: string;
  status_type_id: string;
  summary: string | null;
}

export interface GyomuStatusTypeCdtbl {
  description: string | null;
  id: string;
}

export interface GyomuTaskData {
  application_id: string;
  id: string;
  modified_at: Date;
  modified_by: string;
  parameter: string | null;
  parent_task_data_id: string | null;
  task_info_id: string;
}

export interface GyomuTaskDataLog {
  id: string;
  log: string;
  modified_at: Date;
  modified_by: string;
  task_data_id: string;
}

export interface GyomuTaskDataStatus {
  id: string;
  latest_task_instance_id: string;
  modified_at: Date;
  modified_by: string;
  task_data_id: string;
  task_status: string | null;
}

export interface GyomuTaskInfoAccessList {
  account_name: string;
  application_id: string;
  can_access: boolean;
  forbidden: boolean;
  id: string;
  task_info_id: string;
}

export interface GyomuTaskInfoCdtbl {
  application_id: string;
  class_name: string;
  description: string;
  id: string;
  language: string;
  location: string;
  restartable: boolean;
}

export interface GyomuTaskInstance {
  comment: string | null;
  id: string;
  is_done: boolean;
  modified_at: Date;
  modified_by: string;
  parameter: string | null;
  status_info_id: string | null;
  task_data_id: string;
  task_status: string | null;
}

export interface GyomuTaskInstanceSubmitInformation {
  id: string;
  submit_to: string | null;
  task_instance_id: string;
}

export interface GyomuTaskSchedulerConfig {
  application_id: string;
  description: string;
  id: string;
  is_enabled: boolean;
  monitor_parameter: string;
  next_trigger_time: Date;
  service_id: string;
  task_info_id: string;
  task_parameter: string | null;
}

export interface GyomuVariableParameter {
  description: string;
  id: string;
  variable_key: string;
}

export interface DB {
  gyomu_apps_info_cdtbl: GyomuAppsInfoCdtbl;
  gyomu_market_holiday: GyomuMarketHoliday;
  gyomu_milestone_cdtbl: GyomuMilestoneCdtbl;
  gyomu_milestone_daily: GyomuMilestoneDaily;
  gyomu_param_master: GyomuParamMaster;
  gyomu_service_cdtbl: GyomuServiceCdtbl;
  gyomu_service_type_cdtbl: GyomuServiceTypeCdtbl;
  gyomu_status_handler: GyomuStatusHandler;
  gyomu_status_info: GyomuStatusInfo;
  gyomu_status_type_cdtbl: GyomuStatusTypeCdtbl;
  gyomu_task_data: GyomuTaskData;
  gyomu_task_data_log: GyomuTaskDataLog;
  gyomu_task_data_status: GyomuTaskDataStatus;
  gyomu_task_info_access_list: GyomuTaskInfoAccessList;
  gyomu_task_info_cdtbl: GyomuTaskInfoCdtbl;
  gyomu_task_instance: GyomuTaskInstance;
  gyomu_task_instance_submit_information: GyomuTaskInstanceSubmitInformation;
  gyomu_task_scheduler_config: GyomuTaskSchedulerConfig;
  gyomu_variable_parameter: GyomuVariableParameter;
}